Support hall-of-fame games by allowing the week to be 0.
A better constraint would check both the week and the phase
at the same time, but let's take baby steps.

(I had briefly considered adding a new phase for the HOF game
because having a "0" week bothers me. However, Wikipedia says
the HOF game is officially the start of the NFL preseason.
We shall live with the "0" week: fighting against the NFL's
crazy schedule is already hard enough.)