Treat the sorting interaction between Play and PlayPlayer specially.
This is a pretty bad hack, but basically, we don't want the
play_player table involved in constraining results if we're sorting
by fields that are in the `play` table. Namely, we only want
the play_player table involved if we're sorting by a field that is
in play_player but NOT in play.

That this logic is buried inside the `Sorter` class is a terrible
piece of technical debt, however, the query module is due to be
scrapped and rewritten.